arm: Allow disabling PIE for dynamically linked executables
PIE can and does break some binaries that were built without it,
resulting in relocation-related issues (addressing memory
wrongly assumed to be at a fixed position).

Use this as a last resort measure for compatibility with pre-JB
binary blobs, and only if confirmed to fix a broken executable.
If you're not sure, do NOT enable this on a device.

Change-Id: Ie14757e7f6d7c8a42eadf12a3d1c5eee43536994